+++ app/src/main/java/cn/wthee/pcrtool/data/enums/EnumType.kt
@@ -85,7 +85,8 @@ enum class ToolMenuType(val id: Int) {
     ALL_SKILL(13),
     ALL_EQUIP(14),
     MOCK_GACHA(15),
-    BIRTHDAY(16);
+    BIRTHDAY(16),
+    CALENDAR_EVENT(17);
 
     companion object {
         fun getByValue(value: Int) = values()
