+++ app/src/main/java/cn/wthee/pcrtool/ui/home/Overview.kt
@@ -19,7 +19,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
@@ -34,12 +33,11 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import cn.wthee.pcrtool.R
 import cn.wthee.pcrtool.data.db.entity.NewsTable
 import cn.wthee.pcrtool.data.db.view.*
-import cn.wthee.pcrtool.data.db.view.CalendarEventData
 import cn.wthee.pcrtool.data.enums.EventType
 import cn.wthee.pcrtool.data.enums.MainIconType
 import cn.wthee.pcrtool.data.enums.OverviewType
 import cn.wthee.pcrtool.database.DatabaseUpdater
-import cn.wthee.pcrtool.database.getRegion
+import cn.wthee.pcrtool.ui.MainActivity
 import cn.wthee.pcrtool.ui.MainActivity.Companion.animOnFlag
 import cn.wthee.pcrtool.ui.MainActivity.Companion.navViewModel
 import cn.wthee.pcrtool.ui.NavActions
@@ -51,8 +49,6 @@ import cn.wthee.pcrtool.utils.*
 import cn.wthee.pcrtool.viewmodel.NoticeViewModel
 import cn.wthee.pcrtool.viewmodel.OverviewViewModel
 import com.google.accompanist.flowlayout.FlowColumn
-import com.google.accompanist.flowlayout.FlowCrossAxisAlignment
-import com.google.accompanist.flowlayout.FlowRow
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.HorizontalPager
 import com.google.accompanist.pager.rememberPagerState
@@ -249,10 +245,9 @@ private fun NewsSection(
     overviewViewModel: OverviewViewModel = hiltViewModel()
 ) {
     val id = OverviewType.NEWS.id
-    val region = getRegion()
     //公告列表
     val newsList =
-        overviewViewModel.getNewsOverview(region).collectAsState(initial = arrayListOf()).value
+        overviewViewModel.getNewsOverview().collectAsState(initial = arrayListOf()).value
 
     Section(
         id = id,
@@ -529,7 +524,7 @@ private fun CalendarEventLayout(
         if (isEditMode && !isNotEmpty)
             R.string.no_in_progress_data
         else
-            R.string.tool_calendar
+            R.string.tool_calendar_inprogress
     } else {
         if (isEditMode && !isNotEmpty)
             R.string.no_coming_soon_data
@@ -601,8 +596,7 @@ private fun CalendarEventOperation(
     gachaList: List<GachaInfo>,
 ) {
     val context = LocalContext.current
-    val region = getRegion()
-    val regionName = when (region) {
+    val regionName = when (MainActivity.regionType) {
         2 -> stringResource(id = R.string.db_cn)
         3 -> stringResource(id = R.string.db_tw)
         4 -> stringResource(id = R.string.db_jp)
@@ -683,10 +677,10 @@ private fun CalendarEventOperation(
                         //掉落活动
                         var eventText = ""
                         eventList.forEach {
-                            val date = fixJpTime(it.startTime.formatTime, region).substring(
+                            val date = fixJpTime(it.startTime.formatTime).substring(
                                 0,
                                 10
-                            ) + " ~ " + fixJpTime(it.endTime.formatTime, region).substring(0, 10)
+                            ) + " ~ " + fixJpTime(it.endTime.formatTime).substring(0, 10)
                             eventText += "• $date\n${getTypeDataToString(it)}\n"
                         }
                         if (eventText != "") {
@@ -696,10 +690,10 @@ private fun CalendarEventOperation(
                         //剧情活动
                         var storyText = ""
                         storyEventList.forEach {
-                            val date = fixJpTime(it.startTime.formatTime, region).substring(
+                            val date = fixJpTime(it.startTime.formatTime).substring(
                                 0,
                                 10
-                            ) + " ~ " + fixJpTime(it.endTime.formatTime, region).substring(0, 10)
+                            ) + " ~ " + fixJpTime(it.endTime.formatTime).substring(0, 10)
                             storyText += "• $date\n${it.getEventTitle()}"
                         }
                         if (storyText != "") {
@@ -709,10 +703,10 @@ private fun CalendarEventOperation(
                         //卡池
                         var gachaText = ""
                         gachaList.forEach {
-                            val date = fixJpTime(it.startTime.formatTime, region).substring(
+                            val date = fixJpTime(it.startTime.formatTime).substring(
                                 0,
                                 10
-                            ) + " ~ " + fixJpTime(it.endTime.formatTime, region).substring(0, 10)
+                            ) + " ~ " + fixJpTime(it.endTime.formatTime).substring(0, 10)
                             gachaText += "• $date\n${it.getDesc()}"
 
                         }
@@ -746,7 +740,7 @@ private fun ChangeDbCompose(
     modifier: Modifier
 ) {
     val context = LocalContext.current
-    val region = getRegion()
+    val region = MainActivity.regionType
     val menuTexts = arrayListOf(
         stringResource(id = R.string.db_cn),
         stringResource(id = R.string.db_tw),
@@ -950,163 +944,6 @@ private fun Section(
         }
     }
 
-}
-
-
-/**
- * 日历信息
- */
-@Composable
-private fun CalendarEventItem(calendar: CalendarEvent) {
-    val regionType = getRegion()
-    val today = getToday()
-    val sd = fixJpTime(calendar.startTime.formatTime, regionType)
-    val ed = fixJpTime(calendar.endTime.formatTime, regionType)
-    val inProgress = isInProgress(today, calendar.startTime, calendar.endTime, regionType)
-    val comingSoon = isComingSoon(today, calendar.startTime, regionType)
-
-    val color = when {
-        inProgress -> {
-            MaterialTheme.colorScheme.primary
-        }
-        comingSoon -> {
-            colorResource(id = R.color.news_system)
-        }
-        else -> {
-            colorResource(id = R.color.color_rank_4_6)
-        }
-    }
-
-    Column(
-        modifier = Modifier.padding(
-            horizontal = Dimen.largePadding,
-            vertical = Dimen.mediumPadding
-        )
-    ) {
-        FlowRow(
-            modifier = Modifier.padding(bottom = Dimen.mediumPadding),
-            crossAxisAlignment = FlowCrossAxisAlignment.Center
-        ) {
-            //开始日期
-            MainTitleText(
-                text = sd.substring(0, 10),
-                backgroundColor = color
-            )
-            //天数
-            MainTitleText(
-                text = ed.days(sd),
-                modifier = Modifier.padding(start = Dimen.smallPadding), backgroundColor = color
-            )
-            //计时
-            Row(
-                modifier = Modifier.padding(start = Dimen.smallPadding),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                if (inProgress) {
-                    IconCompose(
-                        data = MainIconType.TIME_LEFT,
-                        size = Dimen.smallIconSize,
-                        tint = color
-                    )
-                    MainContentText(
-                        text = stringResource(R.string.progressing, ed.dates(today)),
-                        modifier = Modifier.padding(start = Dimen.smallPadding),
-                        textAlign = TextAlign.Start,
-                        color = color
-                    )
-                }
-                if (comingSoon) {
-                    IconCompose(
-                        data = MainIconType.COUNTDOWN,
-                        size = Dimen.smallIconSize,
-                        tint = color
-                    )
-                    MainContentText(
-                        text = stringResource(R.string.coming_soon, sd.dates(today)),
-                        modifier = Modifier.padding(start = Dimen.smallPadding),
-                        textAlign = TextAlign.Start,
-                        color = color
-                    )
-                }
-            }
-        }
-
-        MainCard {
-            Column(modifier = Modifier.padding(Dimen.mediumPadding)) {
-                //内容
-                getTypeData(calendar).forEach {
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Subtitle2(text = it.title + it.info)
-                        if (it.multiple != "") {
-                            Subtitle1(
-                                text = it.multiple,
-                                color = it.color,
-                                modifier = Modifier.padding(horizontal = Dimen.smallPadding)
-                            )
-                        }
-                    }
-                }
-                //结束日期
-                CaptionText(text = ed, modifier = Modifier.fillMaxWidth())
-            }
-        }
-    }
-
-}
-
-/**
- * 获取事项信息
- */
-@Composable
-private fun getTypeData(data: CalendarEvent): ArrayList<CalendarEventData> {
-    val events = arrayListOf<CalendarEventData>()
-    if (data.type != "1") {
-        //正常活动
-        val list = data.type.intArrayList
-        list.forEach { type ->
-            val title = when (type) {
-                31, 41 -> stringResource(id = R.string.normal)
-                32, 42 -> stringResource(id = R.string.hard)
-                39, 49 -> stringResource(id = R.string.very_hard)
-                34 -> stringResource(id = R.string.explore)
-                37 -> stringResource(id = R.string.shrine)
-                38 -> stringResource(id = R.string.temple)
-                45 -> stringResource(id = R.string.dungeon)
-                else -> ""
-            }
-
-            val dropMumColor = when (data.getFixedValue()) {
-                1.5f, 2.0f -> colorResource(id = R.color.color_rank_7_10)
-                3f -> colorResource(id = R.color.color_rank_18_20)
-                4f -> colorResource(id = R.color.color_rank_21_23)
-                else -> MaterialTheme.colorScheme.primary
-            }
-            val multiple = data.getFixedValue()
-            events.add(
-                CalendarEventData(
-                    title,
-                    (if ((multiple * 10).toInt() % 10 == 0) {
-                        multiple.toInt().toString()
-                    } else {
-                        multiple.toString()
-                    }) + "倍",
-                    stringResource(id = if (type > 40) R.string.mana else R.string.drop),
-                    dropMumColor
-                )
-            )
-        }
-    } else {
-        //露娜塔
-        events.add(
-            CalendarEventData(
-                stringResource(id = R.string.tower),
-                "",
-                ""
-            )
-        )
-    }
-
-    return events
 }
 
 /**
