+++ app/src/main/java/cn/wthee/pcrtool/viewmodel/EventViewModel.kt
@@ -2,7 +2,9 @@ package cn.wthee.pcrtool.viewmodel
 
 import androidx.lifecycle.ViewModel
 import cn.wthee.pcrtool.data.db.repository.EventRepository
+import cn.wthee.pcrtool.utils.compareEvent
 import cn.wthee.pcrtool.utils.compareStoryEvent
+import cn.wthee.pcrtool.utils.getToday
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
@@ -20,14 +22,13 @@ class EventViewModel @Inject constructor(
     /**
      * 获取活动记录
      */
-    fun getEventHistory() = flow {
+    fun getStoryEventHistory() = flow {
         try {
             emit(eventRepository.getAllEvents(Int.MAX_VALUE).sortedWith(compareStoryEvent()))
         } catch (e: Exception) {
 
         }
     }
-
 
     /**
      * 获取免费十连活动记录
@@ -38,5 +39,20 @@ class EventViewModel @Inject constructor(
         } catch (e: Exception) {
 
         }
+    }
+
+    /**
+     * 获取活动列表
+     */
+    fun getCalendarEventList() = flow {
+        try {
+            val data = eventRepository.getDropEvent() + eventRepository.getTowerEvent(1)
+            emit(
+                data.sortedWith(compareEvent(getToday()))
+            )
+        } catch (e: Exception) {
+
+        }
+
     }
 }
